update windows builder to trixie - change download directory
The windows builder now runs trixie. I also changed the download
directory name to yy-mm-dd-version-gitcommit. I modified the
build process such that the sandbox is download in the permanent
machine upfront. The builder will then only compile. Only when
a tarball is not there, the download will be done during the build.